requirements.txt.txt
tk
requests
numpy
pandas
Pillow
pyautogui
opencv-python
